requirements.txt
flask
flask_sqlalchemy
flask_migrate
psycopg2-binary
python-dotenv
flask_login